+++ R/package-init.R
@@ -1,3 +1,3 @@
-#' @useDynLib runner
+#' @useDynLib runner2
 #' @importFrom Rcpp sourceCpp
 NULL