+++ src/RcppExports.cpp
@@ -12,7 +12,7 @@ Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 
 // fill_run
 SEXP fill_run(SEXP x, bool run_for_first, bool only_within);
-RcppExport SEXP _runner_fill_run(SEXP xSEXP, SEXP run_for_firstSEXP, SEXP only_withinSEXP) {
+RcppExport SEXP _runner2_fill_run(SEXP xSEXP, SEXP run_for_firstSEXP, SEXP only_withinSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -25,7 +25,7 @@ END_RCPP
 }
 // lag_run
 SEXP lag_run(SEXP x, IntegerVector lag, IntegerVector idx, bool nearest);
-RcppExport SEXP _runner_lag_run(SEXP xSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP nearestSEXP) {
+RcppExport SEXP _runner2_lag_run(SEXP xSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP nearestSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -39,7 +39,7 @@ END_RCPP
 }
 // length_run
 IntegerVector length_run(IntegerVector k, IntegerVector lag, IntegerVector idx);
-RcppExport SEXP _runner_length_run(SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP) {
+RcppExport SEXP _runner2_length_run(SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -52,7 +52,7 @@ END_RCPP
 }
 // minmax_run
 Rcpp::NumericVector minmax_run(Rcpp::NumericVector const& x, std::string metric, bool na_rm);
-RcppExport SEXP _runner_minmax_run(SEXP xSEXP, SEXP metricSEXP, SEXP na_rmSEXP) {
+RcppExport SEXP _runner2_minmax_run(SEXP xSEXP, SEXP metricSEXP, SEXP na_rmSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -65,7 +65,7 @@ END_RCPP
 }
 // sum_run
 Rcpp::NumericVector sum_run(NumericVector x, IntegerVector k, IntegerVector lag, IntegerVector idx, IntegerVector at, bool na_rm, bool na_pad);
-RcppExport SEXP _runner_sum_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
+RcppExport SEXP _runner2_sum_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -82,7 +82,7 @@ END_RCPP
 }
 // mean_run
 NumericVector mean_run(NumericVector x, IntegerVector k, IntegerVector lag, IntegerVector idx, IntegerVector at, bool na_rm, bool na_pad);
-RcppExport SEXP _runner_mean_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
+RcppExport SEXP _runner2_mean_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -99,7 +99,7 @@ END_RCPP
 }
 // max_run
 NumericVector max_run(Rcpp::NumericVector x, Rcpp::IntegerVector k, Rcpp::IntegerVector lag, Rcpp::IntegerVector idx, Rcpp::IntegerVector at, bool na_rm, bool na_pad);
-RcppExport SEXP _runner_max_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
+RcppExport SEXP _runner2_max_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -116,7 +116,7 @@ END_RCPP
 }
 // min_run
 NumericVector min_run(NumericVector x, IntegerVector k, IntegerVector lag, IntegerVector idx, IntegerVector at, bool na_rm, bool na_pad);
-RcppExport SEXP _runner_min_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
+RcppExport SEXP _runner2_min_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -133,7 +133,7 @@ END_RCPP
 }
 // streak_run
 IntegerVector streak_run(SEXP x, Rcpp::IntegerVector k, Rcpp::IntegerVector lag, Rcpp::IntegerVector idx, Rcpp::IntegerVector at, bool na_rm, bool na_pad);
-RcppExport SEXP _runner_streak_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
+RcppExport SEXP _runner2_streak_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -150,7 +150,7 @@ END_RCPP
 }
 // which_run
 Rcpp::IntegerVector which_run(LogicalVector x, IntegerVector k, IntegerVector lag, IntegerVector idx, IntegerVector at, std::string which, bool na_rm, bool na_pad);
-RcppExport SEXP _runner_which_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP whichSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
+RcppExport SEXP _runner2_which_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP whichSEXP, SEXP na_rmSEXP, SEXP na_padSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -168,7 +168,7 @@ END_RCPP
 }
 // window_run
 SEXP window_run(SEXP x, IntegerVector k, IntegerVector lag, IntegerVector idx, IntegerVector at, bool na_pad);
-RcppExport SEXP _runner_window_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_padSEXP) {
+RcppExport SEXP _runner2_window_run(SEXP xSEXP, SEXP kSEXP, SEXP lagSEXP, SEXP idxSEXP, SEXP atSEXP, SEXP na_padSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -184,21 +184,21 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_runner_fill_run", (DL_FUNC) &_runner_fill_run, 3},
-    {"_runner_lag_run", (DL_FUNC) &_runner_lag_run, 4},
-    {"_runner_length_run", (DL_FUNC) &_runner_length_run, 3},
-    {"_runner_minmax_run", (DL_FUNC) &_runner_minmax_run, 3},
-    {"_runner_sum_run", (DL_FUNC) &_runner_sum_run, 7},
-    {"_runner_mean_run", (DL_FUNC) &_runner_mean_run, 7},
-    {"_runner_max_run", (DL_FUNC) &_runner_max_run, 7},
-    {"_runner_min_run", (DL_FUNC) &_runner_min_run, 7},
-    {"_runner_streak_run", (DL_FUNC) &_runner_streak_run, 7},
-    {"_runner_which_run", (DL_FUNC) &_runner_which_run, 8},
-    {"_runner_window_run", (DL_FUNC) &_runner_window_run, 6},
+    {"_runner2_fill_run", (DL_FUNC) &_runner2_fill_run, 3},
+    {"_runner2_lag_run", (DL_FUNC) &_runner2_lag_run, 4},
+    {"_runner2_length_run", (DL_FUNC) &_runner2_length_run, 3},
+    {"_runner2_minmax_run", (DL_FUNC) &_runner2_minmax_run, 3},
+    {"_runner2_sum_run", (DL_FUNC) &_runner2_sum_run, 7},
+    {"_runner2_mean_run", (DL_FUNC) &_runner2_mean_run, 7},
+    {"_runner2_max_run", (DL_FUNC) &_runner2_max_run, 7},
+    {"_runner2_min_run", (DL_FUNC) &_runner2_min_run, 7},
+    {"_runner2_streak_run", (DL_FUNC) &_runner2_streak_run, 7},
+    {"_runner2_which_run", (DL_FUNC) &_runner2_which_run, 8},
+    {"_runner2_window_run", (DL_FUNC) &_runner2_window_run, 6},
     {NULL, NULL, 0}
 };
 
-RcppExport void R_init_runner(DllInfo *dll) {
+RcppExport void R_init_runner2(DllInfo *dll) {
     R_registerRoutines(dll, NULL, CallEntries, NULL, NULL);
     R_useDynamicSymbols(dll, FALSE);
 }